Added options for manual sequences and verbose=False
Paired programming with Skye. Realized we needed to add an option to manually input sequences in addition to the file input method. Also added a user input to change "verbose" to false in order to clean up the output for large files. Anywhere a full sequence would be displayed, suppressed the display if verbose is false. Added formatted display of matches and sequences for the unshifted output. Added some more line breaks to help format output.
 Fixed a bug in maxshift user input. Fixed a bug that occurred when sequences are different lengths.